+++ src/main/kotlin/com/rsredsq/revolut/backend/domain/Account.kt
@@ -1,3 +1,3 @@
 package com.rsredsq.revolut.backend.domain
 
-data class Account(val id: Int, val balance: Double)
+data class Account(val id: Int, val balance: Long)
